[refactor] FlowerColorSearch.vue 花の色で、検索するコンポーネントを、作成した。
